Crate oak_session

Re-exports

pub use session::ClientSession;

pub use session::ServerSession;

pub use session::Session;

Modules

attestation

This module provides an implementation of the Attestation Provider, which handles remote attestation between two parties.

config

dice_attestation

This module provides an implementation of the Attestation Verifier bases on verifying the DICE chain.

encryptors

This module provides an interface and an implementation of the Encryptor, which handles encrypted communication over a channel.

handshake

This module provides an implementation of the Handshaker, which handles cryptographic handshake and secure session creation.

key_extractor

Trait for extracting a session binding key from the attestation result to use for verifying session bindings.

session

This module provides an SDK for creating secure attested sessions between two parties.

session_binding

This module provides traits that allow to bind the data to the session.

Traits

ProtocolEngine

Trait that represents a state-machine for protocol message generation. Incoming and outgoing messages are represented as generic arguments I and O.