Trait oak_session::session::Session

pub trait Session {
    // Required methods
    fn is_open(&self) -> bool;
    fn write(&mut self, plaintext: &PlaintextMessage) -> Result<()>;
    fn read(&mut self) -> Result<Option<PlaintextMessage>>;
}

Trait that represents an end-to-end encrypted bidirectional streaming session between two peers.

If one of the methods returns an error, it means that there was a protocol error and the session needs to be restarted (because the state-machine is in an incorrect state).

Required Methods

fn is_open(&self) -> bool

Checks whether session is ready to send and receive encrypted messages. Session becomes ready once remote attestation and crypto handshake have been successfully finished.

fn write(&mut self, plaintext: &PlaintextMessage) -> Result<()>

Encrypts plaintext and send it to the peer.

This function can be called multiple times in a row, which will result in multiple outgoing protocol messages being created.

fn read(&mut self) -> Result<Option<PlaintextMessage>>

Reads an encrypted message from the peer and decrypt it.

This function can be called multiple times in a row, if the peer sent multiple protocol messages that need to be decrypted.

Method returns Result<Option<()>> with the corresponding outcomes:

Ok(None): Nothing to read

Ok(Some(Vec<u8>)): Successfully read plaintext bytes

Err: Protocol error